Last Fixes
Here are my final adjustments to these files. For the views, I added some bootstrap classes to make the tables more visually pleasing. The edits to the other files are to improve the design of the Exercise class. As a result, the LabController's save method now requires Exercise IDs passed in. The exercise IDs are the same as before.